+++ 02-Back-End/webapp/src/main/java/com/S2O/webapp/dao/YearRepository.java
@@ -6,5 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface YearRepository extends JpaRepository<Year, Integer> {
+
+public interface YearRepository extends JpaRepository<Year, Long> {
+
+
 }
